done 2 options from remove menu
