+++ 01_Exercises/Lab 4 to 6/Testing.py
@@ -1,0 +1,4 @@
+# Just testing, nothing to see here :))
+# random_expenses = [["02-06-2022", "20", "Phone"], ["02-06-2050", "55", "Bills"], ["02-06-2022", "20", "Phone"], ["02-06-2022", "20", "Phone"], ["02-06-2022", "20", "Phone"], ["02-06-2027", "20", "Phone"]]
+# the_date_of_expense = "02-06-2022"
+
