+++ 01_Exercises/Lab 4 to 6/removeExpenses.py
@@ -1,0 +1,9 @@
+# Functions for the submenu - Remove Expenses
+def remove_expense(all_expenses, requested_format):
+    updated_expense = []
+
+    for expense in all_expenses:
+        if requested_format not in expense:
+            updated_expense.append(expense)
+
+    return updated_expense
