+++ 01_Exercises/Lab 4 to 6/repository.py
@@ -131,3 +131,9 @@ class Expense:
     def get_all_expense_details(self):
         """It returns all the data together."""
         return (self.date + " " + str(self.amount) + " " + self.expense_type).split()
+
+
+def check_all_data(data_collected):
+    if data_collected.check_date_format() and data_collected.check_amount() and data_collected.check_type():
+        if data_collected.check_day_format() and data_collected.check_month_format() and data_collected.check_year_format() and data_collected.check_amount_int():
+            return True
