--- 01_Exercises/Testing.py
@@ -1,12 +1,0 @@
-# Just testing, nothing to see here :))
-random_expenses = [["02-06-2022", "20", "Phone"], ["02-06-2022", "55", "Bills"]]
-
-def find_expense_location(expenses):
-    location = 0
-
-    for i in range(len(random_expenses)):
-        if random_user_input == random_expenses[i]:
-            location = i
-
-    return location
-
