--- 01_Exercises/funcs.py
@@ -1,2 +1,0 @@
-
-
